Removed `size` property from tokens.
This doesn't make much sense at this moment as we don't do validation
against the JSON ABI from a compiled contract. In future, though, we
might want to reintroduce this feature and check if the size we're
trying to encode (i.e. u64) is less or equal to the size of uint in the
ABI.