Keep a global Secp256k1 context.
This fixes pretty obvious performance bottleneck. Everytime we created
secp256k1 context it was computing a lot of data which turns out to be
really ineffective. Benchmark says its *huge* improvement of astonishing
99.4%.

    sign tx without network id
                            time:   [84.839 us 85.273 us 85.801 us]
                            change: [-99.412% -99.403% -99.393%] (p = 0.00 < 0.05)
                            Performance has improved.

This change reuses Secp256k context just in sign method, but it will be
expanded to other methods as well.

Global context is kept in a thread local storage for performance reasons.