Soft fail on TransactionTests dir check
For some reason the is_dir() function doesn't work correctly on
specific cross architectures. It doesn't impede the tests actually
running so I'm just going to make this check optional. If there's an
actual failure at least the reason will be in the scrollback if not
right in front of the user.