Improve protections for signature derivation
While I don't think this function is used much externally it's still
an API breaking change requiring a minor version bump. All that's
really been done is that I've moved what minor signature correctness
checking we have to the lowest level function interacting with the data.

The fact that this wasn't done in the first place was a bit of an oversight
on my part.