Add support for nested abi encoding
This patch adds support for nexted encoding of structs as function
arguments as outlined in the ABIv2 spec, this comes along with support
for nested dynamic types like multidimensional arrays, although that's
not well tested in this patch.

A new token type 'Struct' is introduced that represents a struct that
can be encoded as a function argument.