Fix encoding for functions with a single struct arg
This patch resolves a strange edgecase where in functions that have one
and only one argument that is also a struct do not seem to need dynamic
offsets. This makes some level of sense as the only data is the dynamic
data. That being said I don' thave a huge amount of confidence in this
change, as I don't have a good backing for it from looking at the spec
document.