Fix encoding hack for arrays of static structs
Previously we had a check for small offset sizes here. This was a hack
created by observing byte dumps.

This patch actually resolves the core issue in a more permenant way.
The core issue is that we treat structs as a dynamic length Token
despite the fact that this is not always true.

Structs with all static members are themselves static in length. By
detecting this we avoid inserting the length parameter and ensure
correctness.

This patch also makes the debug_print_data function public and fixes
some issues around argument coutning when the function signature ends in
[]