Validated PrivateKey
Due to our representation of PrivateKeys as simply an array of bytes
consumers of the Clarity library have to handle errors about the
validaity of their PrivateKey for signing and general use after
constructing the object.

This results in a lot of unwrap statements, the user knows the key they
have is valid and doesn't want to handle the potential that it's not at
runtime.

Since there isn't much valid reason to construct an invalid private key
this library modifies the private key type to perform this validation on
construction, both reducing the number of extra operations to get the
address by caching it and ensuring validity.