Import updated U256
As noted in the num256_rs repo bnum was used incorrectly and specified a
16384 bit integer as opposed to a 256 bit one. This upgrade imports the
fixed version of num256 and removes some now redundant overflow checks.

Previously it was not notied that tt256 actually should overflow a 256
bit integer since we moved from one integer library with a larger
internal width to another. Now that we have proper range checking at the
foundational level we can remove some checks and fix our constants.