Fixes for EIP1559 tx
This updates both clarity and web30 to resolve issue with submitting
eip1559 transactions. On the clarity side the single byte v in new
format signatures needs to be formated as a singel byte rlp token.

On the web30 side we must actually attach the signatures to our
transactions if we want them to go through.