Split prepare and tx publish
This patch splits the prepare and tx publish roles. For a long time
we've had send_transaction fully abstract the lower level details of
making a payment. But this doesn't cover possible cases where a
transaction has been handed off to a full node, and could potentially
not be published.

In order to resolve these cases robustly we leave it on the user to do
final publishing, having access to the transaction object the caller can
retireve the txhash and perform some informed queries to discover if it
was actually published despite the full node returning an error.