WIP
Lots of modifications made to the ethers-rs code, but it is looking like
this is still a huge amount of work to swap out the core of the language
parsing and translation in a way that will work for us long term. Maybe
this will be useful in the future but at the moment it is still daunting
and a lot of code to copy in and replace. We have the goal of replacing
the use of the ethabi crate as well, so that is a significant amount of
work.