Remove net_version and use eth_chainID instead
The net_version endpoint is not as strictly defined as the eth_chainID
endpoint as it's not officially part of the jsonrpc spec. As such it can
sometimes disagree with the net_version endpoint in some situations (one
of which I've just encountered and motiviated this change).

This was made even more confusing to the user by the fact that the
automatic transaction builder used net_version.

This patch removes the net version endpoint, which was mostly used
becuase it returned values in decimal instead of hex. The eth_chainID
endpoint returns hex, which can be implicitly decoded by handlers we
setup for Uint256, then downcast. The previous version of the endpoint